requirements.txt
requests
yt-dlp
requests
python_dotenv
SpeechRecognition
flask
google-generativeai
json-repair
pillow
lxml
faust-cchardet
beautifulsoup4